Absorb LeaderScheduler::get_active_set()
No functional changes
